Keep hosted view lifecycle test within file budget
